+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserService.java
@@ -8,13 +8,13 @@ import java.util.List;
 
 public interface UserService extends UserDetailsService {
 
-    User getById(int id);
+    User getById(Long id);
 
     List<User> listUsers();
 
     void saveUser(User user);
 
-    void removeUser(int id);
+    void removeUser(Long id);
 
     void updateUser(User user);
 }
